Ventana update small fixes, ready for now
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@35102 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: c251a6cd4a71faca1013f35bb277796a9e1f18de

+++ src/router/rc/update-after-pivot.sh
@@ -1,4 +1,15 @@
 #!/bin/sh
+if [ x$1 = x ]
+then
+	echo "No file/fifo given, exit"
+	exit
+fi
+if [ x$2 = x ]
+then
+	echo "No mtd partition given, exit"
+	exit
+fi
+
 FIFO=$1
 MTDPART=$2
 
@@ -14,4 +25,7 @@ mtd -f write ${FIFO} ${MTDPART}
 busybox sync
 busybox sync
 busybox sync
-busybox reboot
+if [ x$3 = x1 ]
+then
+	busybox reboot
+fi
